+++ src/main/java/com/cusbee/salers/controller/UpworkBidController.java
@@ -27,12 +27,12 @@ public class UpworkBidController {
     }
 
     @RequestMapping(value = "/delete", method = RequestMethod.DELETE)
-    public void delete(@RequestParam(value = "id") Integer id) {
+    public void delete(@RequestParam(value = "id") Long id) {
         upworkBidService.delete(id);
     }
 
     @RequestMapping(value = "/changeActivity", method = RequestMethod.POST)
-    public void changeActivity(@RequestParam(value = "id") Integer id) {
+    public void changeActivity(@RequestParam(value = "id") Long id) {
         upworkBidService.changeActivity(id);
     }
 
@@ -52,7 +52,7 @@ public class UpworkBidController {
     }
 
     @RequestMapping(value = "/getBySaler", method = RequestMethod.GET)
-    public List<UpworkBidBean> getBidsBySaler(@RequestParam(value = "salerId") Integer salerId) {
+    public List<UpworkBidBean> getBidsBySaler(@RequestParam(value = "salerId") Long salerId) {
         return upworkBidService.getBySaler(salerId);
     }
 
@@ -67,13 +67,13 @@ public class UpworkBidController {
     }
 
     @RequestMapping(value = "/getBySalerAndDateBetween", method = RequestMethod.GET)
-    public List<UpworkBidBean> getBidsBySalerAndDateBetween(@RequestParam(value = "salerId") Integer salerId,
+    public List<UpworkBidBean> getBidsBySalerAndDateBetween(@RequestParam(value = "salerId") Long salerId,
             @RequestParam(value = "dateFrom") String dateFrom, @RequestParam(value = "dateTo") String dateTo) {
         return upworkBidService.getBySalerAndDateBetween(salerId, dateFrom, dateTo);
     }
 
     @RequestMapping(value = "/getBySalerAndUpdatedDateBetween", method = RequestMethod.GET)
-    public List<UpworkBidBean> getBidsBySalerAndUpdatedDateBetween(@RequestParam(value = "salerId") Integer salerId,
+    public List<UpworkBidBean> getBidsBySalerAndUpdatedDateBetween(@RequestParam(value = "salerId") Long salerId,
             @RequestParam(value = "dateFrom") String dateFrom, @RequestParam(value = "dateTo") String dateTo) {
         return upworkBidService.getBySalerAndUpdatedDateBetween(salerId, dateFrom, dateTo);
     }
